running_mate_strava_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
kafka-python